Don't miss the real zero value when exporting
If one kind (e.g. cadence) of data is missing, all values are zeros for it, thus no need to to check for single values for export. If sometimes zero value is real data (e.g. temperature in winter) - it will not be skipped because we know that same array has also non-zero data.